Finished simulation Thm, added some explanation before and after
Added Proposition 1 (proof is simple, so can be moved to appendix). Finished proof of Theorem about simulation. Changed statement from 2D Turing machines to algorithms, because otherwise we may need to add many unnecessary details about how the input is given on the 2D tape, how we traverse it, etc.