Added unary, arity 1, delta 2 case +Figures
I added the proof for delta 2, arity 1, unary. It needs some polishing still, but the main points are in there.
I also added a Fig folder with svg and png figures in case it is useful.
For the same reason, uploaded the old manuscript which has the write up for my version of the Chichen Itza construction, in case it helps Reoto