Visual editor for Oritatami written in Processing.py
The last couple of days I was working on this program to have a unified look for the diagrams. It can visualize conformations and one can edit and save them.
Even if we do not use it for this project, it will be useful later. When I will have time, I will add the algorithmic parts: folding, etc.

To run it, you need to download Processing from processing.org and install the python mode.

I will add documentation in the days to come.