Added pyramid part, polished arity 1, delay 3, unary
The pyramid part still needs some work, as does the arity 1, delay 2.

I went over Reoto's arity 1, delay 3 and made some changes to make it more readable. For now I commented out the last 2 paragraphs, as they are not necessary for the proof, I think. However, here we need to add some conclusion to the arguments of the previous two subsections.